Load completions after bundles are loaded.
bundle-init is now renamed to bundle-apply and is to called after all bundle
specifications are given. Also added notes in the README file.

The reason is the compinit function needs to be run after adding all the
required directories to `fpath`. The only reliable and performant way to do it
would be to explicitly call a function that does it after specifying all the
bundles.