Fix current theme is listed in cleanup.
Current theme and the plugins installed on-spot in current session should not be
listed in the cleanup list. They were because of a problem with updating the
`$bundles` record. This is now fixed.

If the above `if` is directly piped to the below `while`, the contents inside
the `if` construct are run in a new subshell, so changes to the `$bundles`
variable are lost after the `if` construct finishes. So, we need the temporary
`$install_bundles` variable.