Do comparisons with directory names in -cleanup.
The directory names were being converted to urls and then compared to the
contents of the record. The problem with this is there could be directories with
extremely wierd names in there, not created by antigen. This could throw the
dir->url conversion off.

So, with this, the urls from the record are converted to directory names and are
compared with the contents of repos/. Any directory unaccounted for will be
converted to a repo url, solely for the purpose of displaying to the user. After
confirmation, these directories are deleted, whether they represent a git clone
or not.