Memoization is not possible in a sane way.
The memoized results of these functions cannot be saved in a global hash map,
becuase these functions are always run in a subshell. Changes to variables in
the subshell do not reflect in the main shell instance. We'd have to resort to
use the `$REPLY` ways returning the result from these functions, if they were to
be memoized. But that way of doing things sucks. Totally unreadable. So, just
not memoizing.