Use './' instead of $PWD to find current dir.
@desyncr it probably will work now with `sudo`...most likely `$PWD` was
set as an environment variable (that's where I got the idea from) on
your own login shell, but sudo does not (IIRC) run as a login shell, so
it might've not been set on your system. Either way, this change removes
the call to `$PWD` and uses './' as the prefix to antigen.zsh to
communicate we want the file in the current dir.

Also...you shouldn't need that extra assignment of `PYENV=` in the
command, since `make` just runs clean and install...you can run `make
tests` separately to run tests.