Update antigen completion
As discussed on [#26](https://github.com/zsh-users/antigen/issues/26), currently antigen completion does not include subcommand arguments.
So I tried to update it. But since first writing the zsh completion, it may be wrong, and aruguments may not be all included (I have not
seen the whole source code yet). So that, I wip this PR. Please review this PR, so I fix it.