Extension system for antigen
Provides an extension load mechanism for loading custom code that extends antigen.

This is different than plugins or themes in a way that it actually interacts with
antigen as it can hook into different antigen commands and internal/private functions.

It provides pre- and post- hooks for any function (currently even outside antigen). Effectively
executing the pre-function, the function itself and a post-function.

It doesn't short-circuit (meaning it will always call all hooks and the actual command/function).

Usage:

      antigen ext ext-name # loads (ie, sources) the given extension from $ADOTDIR/ext/

These are only meant to be run inside extension definitions:

      # overwrites 'any-antigen-function' (ex, antigen-bundle) with 'extension-hook-function'
      # which could be any function already defined.
      # it can hook into pre- or post- events. defaults to pre-.
      -ext-hook "any-antigen-function" "extension-hook-function" ["pre"|"post"]

      # defined a new autocompletion for the antigen command (ex, antigen logger)
      -ext-compadd "extension-auto-completion"

The hooks functions receives the same arguments the original function receives.