Fix antigen-update if noclobber is set in zsh
When noclobber is set in zsh, the second and all following "antigen
update' may face issues as the revert date is not correctly written to
the file.

As a reminder, noclobber unsets the CLOBBER flag. When the flag is not
set, existing files cannot be truncated when using a simple ">"
redirection, and it needs to be forced with ">!".